Add info about output variables to html display
html_params.xml is a little ADMS backend to create an HTML data page for
Verilog-A models.  It started life just as a way to display all the
parameters a device declared (hence the name) but has grown to something
I use to explore the data tree.

This commit adds a table of all the variables a model declares that the
Verilog-A standard says we're supposed to allow users to print somehow
(either by ".op" or some other mechanism).

I've realized that there are oddnesses in the ADMS data tree that
require one to be pretty careful in ones conditionals before concluding
that a variable is actually an output variable, even if it's marked as
such by ADMS (for example, it's NOT really an output variable if it's
also an INPUT variable (parameter), even though ADMS has marked it as
such somehow).  This little HTML back-end is written to be careful in
that way.

In doing this, I have seen that PSP103 declares 164 such variables.
This could be an issue for us, given that in order to do output
variables by the mechanism Eric created for the BSIM4 and other devices,
we have to create a slot in the store vector for each of them, whether
the user has decided to print them or not.

http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=788