Removal of ground node from Jacobian matrix
The ground node has been represented in the Jacobian graph with an additional
row and column, with entries predicated by device connections to ground.  That
has enabled devices to be agnostic of their attachment to ground.  However,
the addition of the ground node to the Jacobian graph generates overhead for
direct solvers, since they often expect the matrix to be stored contiguously
in memory.  This means, in serial, that the matrix needs to be copied every time
the direct solver is called upon to solve the linear system.  This copy is also
carried out through the Epetra_Export system.  For small matrices, which are the
ones most often used with direct solvers, this overhead is apparent in the
performance of the linear solver layer.

The solution being committed changes the topology layer, which is responsible
for converting the device GIDs to LIDs, such that it returns -1 as the LID and
not a valid LID into the memory of a matrix or vector.  The matrix and vector
classes have been modified to contain a groundNode scalar that they return the
address or pointer to when -1 is used as an index in either the [] or () operators.

Additional machinery to determine the LID of ground nodes was added to the block
analysis layers and linear algebra, which had to be fixed since a GID of -1 is
now an LID of -1.  The builder only generates matrices that use the JACOBIAN
graph and JACOBIAN_OVERLAP graph.  So, in serial, the overlapped matrix is now
the assembled matrix.  In parallel, they are still different matrices and the
assembly is still required.  However, in both serial and parallel, the notion
of the ground node in the graph is gone.