Add alternate ADMS back-end
This is the version of the back-end I used to generate the starting
point for hand-hacked derivatives in VBIC, after having learned a
thing or two doing more hand-hacking in BSIM-CMG.  Instead of
generating Sacado code, it outputs comments to be cut/pasted into
maxima.  I intend to use it as a starting point for really generating
derivatives ourselves.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168