Implement some analytic derivatives for Verilog-A
This is a first, timid step at getting Xyce/ADMS to generate
derivatives without relying on Sacado.

This only works if you aren't using ANY function calls at all, except
for "ddt" (which is a fake-out).

With it, we appear to be processing simple resistors, capacitors, and
RLC devices correctly.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168