Fix some ugly derivatives, add one more
Some special cases where one of the derivatives of a binary
operator's arguments is zero were not being handled specially, and so
made for lots of ugly output like ((0.0+foo)-0.0).  This fixes it.

Also, return "0.0" for the derivative of a string (never really
needed, but sometimes we try to access it, and better get zero).

With this, I can process VBIC with the new back-end, and so long as
expressions don't use function calls, I get things that look like the
hand-generated code.  Of course, VBIC uses function calls all over the
place, so this is nowhere near ready to run on VBIC.  But it doesn't
crash when you try it, and some of the lines look right.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168