Add function precomputation code
When processing assignments such as:
    foo = func1(bar+func2(baz))

and needing to do explicit derivatives (instead of relying on Sacado),
it is necessary to precompute things like func2(baz) and
func1(bar+func2(baz)) so that they can be referred to multiple times
in processing the assignment and its derivatives.  It is also
necessary to compute the derivatives of func1 and func2 with respect
to their arguments, and do chain rules.

This commit makes the first step of pulling function calls out of
assignments and precomputing them, then using the precomuted values in
assignments.  It does this in both normal and sensitivity contexts,
because it's correct to do in both and shouldn't cost much.

No derivatives are being handled yet.  That will require additional
precomputation steps.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168