Start adding function derivative precomputation
When an assignment has function calls on the RHS:

    foo = bar(baz,bletch);

and any of the arguments of the function depend on voltage probes, one
has to evaluate the derivatives of the function with respect to its
arguments, then use the chain rule to get the derivatives of the LHS:

   dfoo_dprobe = dbar_darg1*dbaz_dprobe + dbar_darg2*dbletch_dprobe;

This commit adds the beginning of this.  The way it works is that each
verilog-a assignment that involves function calls causes the single
verilog-a assignment to be emitted as a block instead:

   {
      double value_bar_0 = bar(baz,bletch);
      double deriv_bar_0_d0 = (expression for the derivative of
                               bar(baz,bletch) w.r.t. arg 1);
      double deriv_bar_0_d1 = (expression for the derivative of
                               bar(baz,bletch) w.r.t. arg 2);

      dfoo_dprobe = deriv_bar_0_d0*dbaz_dprobe+deriv_bar_0_d1*dbletch_dprobe;
      foo = value_bar0;
   }

This commit only implements the function derivative for "exp" and
"pow", and does not yet actually do the chain rule part for
dfoo_dprobe yet.  But I wanted to checkpoint the state of the back-end
that actually emits something and doesn't have grotesque errors that
cause it to puke.

So far, it looks like I'm on track to get the "toys/diode2.va" model
so it can be processed this way.  Next step is to get the actual
emission of the foo derivative correct.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168