Add function derivative handling
This commit completes the handling of function derivatives for a tiny
subset of allowed functions, and fixes an error in the output of
devices that use contributions into voltage branches.

With this commit, we handle "pow" and "exp" correctly, evaluating the
function and its derivatives without resorting to Sacado.  This is
sufficient to get the "diode2.va" model in the ADMS/examples/toys
directory built and functioning correctly.

Because the "dump out maxima commands" version of the back-end from
which this was based had only ever been applied to the VBIC model, I
neglected to update the code for computation of the extra static
contributions that result when the Verilog-A input contains
contributions into voltage branches --- these add a branch current
that ADMS itself has no idea about, and have to be handled completely
by our own hacks.

With this commit, I can compile the "toy" resistor, capacitor, diode2,
and vsrc devices into a shared-library plugin and run the
DiodeClipper.cir circuit in the toys directory, producing the correct
output.

I still have a lot of work to do, because all the other one- and two-
argument functions supported by Verilog-A have to be implemented
correctly with their derivatives, but this is a good start.
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168