Add warning when unimplemented function encountered
At this point, we have a working non-sacado back-end that works for
simple models, but has lots of missing capability.  It is not ready to
use on real models like VBIC, though it is close for VBIC (and very
far from close for something like MVSG-HV).

This commit just makes sure we are warned when a model uses some
feature that isn't yet implemented.  It already put comments in the
output saying that there was something missing, but this gets more in
your face about it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168