Fix broken sqrt output and noise output
There was an extra right paren being emitted when sqrt was processed.

The noise contributions have to disable the "substitute variables for
function calls" behavior, because they don't involve precomputation or
derivatives.

With these two changes, we now can process VBIC and it will compile.
And amazingly, the code so generated even runs and produces correct
output.

This version is still a little clunky, though, because it emits code
that causes some annoying compilation warnings.  For example, if one
has a line of code like:

   Ibex = Ibex + <expression that depends on only one of the probes
                  that Ibex already depends on, say V(a,b)>

you wind up with things like
   d_Ibex_dV_c_d = d_Ibex_dV_c_d;
   d_Ibex_dV_e_f = d_Ibex_dV_e_f;
   d_Ibex_dV_a_b = d_Ibex_dV_a_b + <derivative of extra bit>;

That gets you warnings from some compilers about the first two lines
assigning a variable to itself.

A smarter back-end would skip the output altogether if the left and
right hand sides end up identical.  I think I can do that, and it'll
be the next thing I try to clean up before moving on to more glorious
hackage.

Finally, we add a line to the xyceHeaderFile.xml file that makes sure
it won't cause ADMS to puke on the "skipFunctionPrecomp" variable
being undefined when the header XML is listed before the
Implementation XML, or when used alone.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168