Eliminate self-assignments, change error to warning
The code to spew derivative expressions was sometimes emitting things
like:
   d_foo_dbar = d_foo_dbar;

when foo depended on multiple independent variables, but the
particular expression on the RHS of
   foo = foo + <expression>

didn't depend on bar.  This commit prevents ADMS from emitting
anything at all if we're just assigning a derivative value to itself.

I am also changing the UserError to a UserWarning in the cases of
parameters being out of bounds.  We have traditionally treated these
as fatal errors, because the user is giving a parameter out of the
bounds specified by the model author.  However, we have seen some
cases of model cards that violate these recommended constraints, and
we would be better off warning and pressing on than just bombing.
Users who experience simulation problems when using parameters out of
bounds in the face of these warnings are on their own.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168