Set VBIC 1.3 3T verilog attributes for level 11
When I hand-differentiated the VBIC 1.3 3-terminal device, I changed
the verilog-A file so it called the ADMS output "level 111" instead of
level 11, lest one accidentally regenerate it from source and clobber
the optimized level 11.

Now that the new back-end can produce derivatives itself, I'm
restoring the level 11 attributes to this file.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168