Add limexp derivative support
Limexp is a commonly used "limited exponential" function.  In Xyce
implementation (as with many other open source codes), it's just the
exponential below a certain value of the argument, and linear above
that.

Adding support for this to the automatic differentiation allows HICUM
Level 2 to build.  Level 0 does not yet work because that requires
atan and cos, which are not yet implemented in
xyceBasicTemplates_nosac.xml.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168