Get PSP (102 and 103) working with new Xyce/ADMS
This one took more than just adding a few functions to the automated
differentiation, because it exposed holes in the back-end, notably
when function calls appear on the RHS of contributions.

This commit doesn't get them all right, either --- there is still a
gap if a contribution is into a voltage branch and the RHS has
function calls that depend on probes.  But most other cases are handled
correctly (insofar as they are tested now).

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168