Fix backend to avoid generating whitespace error
When the back-end generates statements initializing derivative
variables to zero (which it does when an assignment has an LHS that
can theoretically depend on a variable, but an RHS that does not
depend on that variable) it was emitting an extra space at the end of
the line.  This is something that git can whine about due to its
inclusion of certain Linus Torvalds stylistic hangups.