Move a paren for clarity
In one case, the back-end was generating a ternary operator on the RHS
of a derivative expression that relied on the precedence of the
ternary operator to be correct.  Rather than do that, it is clearer to
make sure to put parens to make explicit where various terms go.

This has absolutely NO consequence on the evaluation of expressions,
but is still a good change because it makes the intent absolutely
clear, and does not rely on the often-confusing precedence rules of
the ternary operator for its correctness.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168