Make extra comments be emitted
There is one special case being handled where the LHS of a
contribution is a potential branch (e.g. "V(sf,gnd)") for which there
are no other contributions into the corresponding flow
branch (e.g. "I(sf,gnd)").  In this case, there is no "probe"
associated with the branch, and therefore we have to emit some code
that directly references nodal values instead of branch voltage drops
via a voltage probe.

In this case, we can't use the usual mechanism of storing the
derivatives, and we never have, even with Sacado.  It is necessary to
add or subtract a one from the relevant Jacobian entry in the
loadDAEdFdX function directly.

Because this operation is far removed from the spot where
contributions are processed, it leaves the developer who is looking
over code and checking for correctness in a position of having to read
way more ADMS back-end code to remember why there's no derivative
being emitted than should be required.

This commit simply makes ADMS emit a comment explaining why there's an
apparently missing derivative.  This is useful both for my purposes of
debugging *AND* to stave off any queries from hypercurious users who
actually read code.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168