Make fabs, min, max derivs compatible with sacado
I had a mistake in the evaluation of the derivative of fabs with
respect to its argument.  That was enough to cause some serious issues
at the first step of one of the MVS 2.0 test suite tests when compared to the
original Sacado version.

The issue is that the derivative of fabs(x) was being evaluated
as (x>0)?1.0:-1.0, whereas Sacado does it as (x>=0)?1.0:-1.0.  A
similar issue with min and max

This is not enough to fix all the problems with MVS, but it needed to
be done.  It may impact other devices, too.
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168