Add support for hyperbolic funcs and inverse funcs
This adds support for cosh, sinh, tanh, their inverses, and the
inverse trigonometric functions acos and asin.

In addition to that, I've made the back end very verbose when it hits
a function it doesn't know how to deal with, lest we see ADMS complete
silently, but completely omit correct differentiation or badly
generated function calls.

I also now fall back to a straight function call if we aren't one of
the small number of functions we are precomputing.  Only functions
that show  up in "rhs/function" get precomputed, and for some
reason (probably a good one) only built-in functions are in that
list.  Analog functions never are.

With this change, we are no longer improperly emiting uses of the
variable "value_<analogfunc>_-1" when we hit an analog function call.
We emit the actual call to the function.

This commit is NOT enough to get models that use analog functions
compiled.  The derivatives of analog functions with respect to probes
are NOT being computed.  But at the moment, we're also emitting "(0.0
/* should have deriv_<analogfunc>*/)" where that derivative belongs,
and try not to do so when the arguments of the analog function don't
depend on the probe we're differentiating with respect to at all.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168