Fix handling of function precomputation
The "assignment" template was unconditionally setting itself up to do
function precomputation *and* to use precomputed values.  This is
wrong if the same template is also to be used to generate assignments
in contexts where that is not needed or desired (as in the generation
of analog function templates).

I have moved the setting of the skipFunctionPrecomp variable out of
the assignment template and into the xyceImplementationFile_nosac.xml
file.  That allows us to exclude chunks of code from this usage.

Right now, both normal context and sensitivity context are getting the
function precomputation, and only normal context gets function
derivative precomputation.  There is really no value to having
function precomputation in sensitivity context, but neither is there
really any great harm.  So I'm leaving it there.  But if we wanted to
turn it off, it's just a matter of moving an "<admst:variable..." line
so that it sets skipFunctionPrecomp to "n" before it emits sensitivity
code.

By doing this, we prevent the incorrect use of function precomputation
in analog function templates.  The handful of analog functions that
were getting processed by this code up to now were simple enough that
this issue wasn't a problem, but it was still wrong.

Only VBIC actually got any code changes out of this.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1168