Silence a reorder warning
C++ requires that all members of a class that are initialized in
constructor initializers be done in the order that they are declared.
If one puts initializers into the constructor out of order, the
compiler reorders them to comply with this requirement.  Under some
circumstances compilers will warn you about this, but not always.

I just corrected a new such issue in the Vsrc that was introduced a
few days ago.  HBSpecified_ and ACSpecified_ were being declared in a
different order than they were initialized.  THe compiler warning was:
../../../../<directory>/src/DeviceModelPKG/OpenModels/N_DEV_Vsrc.C:958:5:
warning: field 'HBSpecified_' will be initialized after field
'ACSpecified_' [-Wreorder]
    HBSpecified_(factory_block.deviceManager_.getHBSpecified()),
        ^

Some of our customers (notably Alegra) compile with "-Wall" because
they consider all warnings to be signs of suspect code.  As a result,
we have occasionally gotten complaints about the number of warnings we
emit, and this particular reordering warning was one I addressed at
their request some time ago.  Unfortunately, they keep coming back
because it's an easy mistake to make.

I ask that developers please add "-Wreorder" to their CXXFLAGS so
that they are more aware of this and correct the warnings when their
new code causes this to be emitted.