Modified embedded sampling to output # quad pts
When I run embedded sampling with quadrature PCE, I realized today that
there was no output that directly stated how many quadrature points were
being used.

This is less of an issue when using regression PCE, or just sampling,
b/c for those methods the user simply specifies the number of sample
points.  So in theory the user should know this.

But for projection (quadrature) PCE, the number of points is determined
from the quadrature integrals, the number of uncertain parameters, the
order of the expansion, whether or not one is using sparse grids, etc.
So it isn't easy for a user to know how many points are being used.

However, knowing it is important, as it tells the user how much larger
the expanded block system matrix will be, and should give a rough idea
how much slower the UQ calculation will compared to the original
problem.