extend the port device for other analyses
The port device is developed to support the S-parameter analysis.
During the calculation of S-parameters, the impedance information
is needed, but it does not connect to the voltage source.

Hspice can also use port device for other analyses. When it is
used for DC and transient analysis, the port device behaves as a
voltage source in series with an impedance. IN Xyce, the port device
is mapped to a voltage source. In order to support other analyses
for port device, the Jacobian and vector loading will be changed.
An internal variable will be added.  In this commit, a compact and
simpler implementation is used. The DAE equation for port device
is formulated in a way such that no internal variable is needed.
Only the last row of Jacobian matrix and F vector loading are changed
when port device is used for other analyses. The extended port
device is tested for DC, AC and transient analysis. They match the
Hspice results