Place a preamble at the top of the generic INSTALL file
The contents of INSTALL was the standard, generic stuff that Autoconf
will create if you have no file called INSTALL.

If a user read this file and tried to install Xyce, they'd be sadly
disappointed that what it says to do doesn't work.  Trying to build
Xyce without following the guidance in the building guide (the vast
majority of which is about getting all the correct TPLs in place
first) is a bad idea, and now INSTALL says so.

I have added some text at the top pointing the user to the building
guide, and then pointing out that the subsequent stuff is generic
documentation of configure.

The INSTALL document has been in place since 2013 and has been this
bad ever since.  Prior to that it had last been touched in 2003 in
commit a20694e0, when it contained what was then detailed guidance for
building Xyce on Xyce team computers, and was intended to be read only
by developers.  That stuff was replaced in 2013 because it was
completely obsolete, and we were getting ready for an open source
release --- as a stopgap, I had just used the generic INSTALL from
autoconf.  It has not been revisited or corrected since then.