Making the rnd number functions toggle on cmd line
AGAUSS, GAUSS and RAND are all Hspice-related functions that are
supported in the expression library.  However, they aren't fully
supported in the sense that they don't connect to the .SAMPLING and
other embedded UQ methods.  That is the subject of a bug that I don't
have time to close for 6.11.

In Hspice, the sampling functions return the mean when MC analysis is
not enabled, and return random numbers distributed on the specified
distribution when sampling is enabled.

The original implementation of AGAUSS, GAUSS, etc just returned the
mean, but they were modified later to return a random distributed
number, on the assumption that the person running Xyce was going to call
it a bunch of times from a script.    However, it was done in a way that
was impossible for the user to turn off.

This comming adds a new keyword command to Pete's parsing manager
-hspice-ext option.  The new command is "random", and it is a boolean
with a default value of false.

If set to "true" then the random-related expression functions will work
as before.  If set to "false" (the default) they return the mean.

In modern PDKs, there are often thousands of instances of GAUSS, etc.
SO it was a significant translation job, in terms of the # of lines of
PDK files that had to be modified.  This change should make it possible
avoid that translation step.   However, for PDK's that are already
translated, this also won't break anything.

There are four tests (that I have found) that this affects:

Certification_Tests/BUG_39_SON/agauss
Certification_Tests/BUG_39_SON/gauss
Certification_Tests/BUG_742/bug_742
Certification_Tests/BUG_877_SON/agauss

these tests will need to be modified to use the new command line option.
In my local copy I have done this and it works.