Fix Use of TEMP in .GLOBAL_PARAM for "Non-Step" Case
The use of TEMP and VT in .GLOBAL_PARAM statements, that are then
used within device instance parameters, was fixed for the .STEP
case for Xyce 6.10.  This commit fixes it for the "non-step" case
by passing in the "current device manager's value of the temperature"
as a parameter to the addGlobalParameter() function.

SON Bug 1033
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1033