Update so it will compile with old breakdown model
As a quick test for bug 1144 son, I compiled the diode with Keith's
"new" breakdown model (from 2007) turned off.  I discovered that because
of the sensitivity analysis code not being type-correct, it wouldn't
compile.

It was very easy to fix to make it compile-able, and that is what this
commit does.  It doesn't fix any current issues, but is the right thing
to do.