Fix spice-incompatible diode discontinuity
In commit edc96c8, a Xyce-specific modification was made to the
standard SPICE3F5 diode's reverse current equation, adding a new
Xyce-specific parameter called "IRF" that was used to multiply the 3F5
diode's reverse current formula.  The default value of this parameter
was 1.0, so the intent was that it should fall back to SPICE3F5
compatibility if the user hadn't specified it, and provide a
user-tunable parameter that could be used to match experimental data.

Later, in commit 1795ae1, a temperature-dependent correction factor
was added to the formula, but no care was taken to make this revert to
1.0 if the user had not specified IRF --- as a result, a
temperature-dependent discontinuity in the diode current was
introduced, making the device inconsistent with the SPICE3F5
implementation at any temperature other than TNOM.

Still later, in commit 2a543e1, the conditional that caused the IRF
factor to be used was modified, introducing an additional temperature
dependence on the location of the discontinuity.

Xyce's level 1 diode has always been intended to be a faithful
implementation of the SPICE3F5 diode, and its level 2 diode was
intended to add a small number of additional equations to improve its
compatibility with PSpice.  The introduction of an unconditional
temperature-corrected IRF factor, and the later introduction of a
temperatature-dependent discontinuity point, was never intentional.

This commit makes the setting of "tIRF" (the factor actually used in
the reverse current formula) completely dependent on whether the user
had specified IRF in the model card --- if not, tIRF never has the
temperature correction factor applied, and the device reverts to the
standard SPICE3F5 formulation as originally intended.

But if the user specifies any value of IRF (even its default value of
1.0), tIRF is IRF*pow((TEMP/TNOM),1.6), thereby always introducing the
potential for a discontinuity in the diode current.  For this reason,
Xyce now emits a warning if the user ever specifies IRF on a model
card for a diode.

IRF is now deprecated, and users are advised not to use it.

Tests and documentation to follow.