Document IRF and deprecation of IRF in diode
The documentation of the IRF parameter in the diode documentation was
incomplete and incorrect --- it failed to document a temperature
correction that was being applied.

Furthermore, it claimed "strict spice compatibility" if IRF was not
set or was set to 1.0, which was a false statement until today --- the
temperature correction had been getting applied irrespective of
whether IRF was being set.  I have changed that today, so if IRF is
not specified we DO have strict spice compatibility of the model.  I
have also documented what happens if IRF *IS* specified, even if it's
set to 1.0 in the specification.

Finally, since the IRF fudge factor was a Xyce-specific modification
of the diode from 2005 and has never been properly validated in any
way, it is a bad idea to encourage users to use it to fit experimental
data.  Bug 1144 SON demontrates that it introduces a discontinuity
that can lead to incorrect results and simulation failures in some
cases.  For this reason, I am declaring the feature deprecated, and it
may be removed in the future.  The documentation now warns users to
stay away from it, as does the code.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1144