Removing commented-out code related to bug 414son
I was the person who fixed bug 414 SON 4 years ago.  The bug was about
fixing some of the Newton solver convergence tests, and the fix included
removing (commenting out) one of those tests.

I left this code in place to be cautious, but at this point (4 years
later) I think it is safe to say that there have been no negative
consequences of its removal.  So, removing completely for real.