Fix bug 1150 SON - Deprecate temperature interp
I have conferred with dept 9434 and they have informed me that they do
not use the temperature interpolation capability.  As it is documented
and tested by roughly 11 tests, I didn't delete it but have added a
warning message which states that this capability is deprecated and will
be removed from a future release.