Add SON Bug 1033 to Xyce 6.11 Release Notes
This commit adds SON Bug 1033 to the Xyce 6.11 Release Notes.
That bug is about the use of TEMP and VT in device instance
parameters and global parameters.  It is being added to the
Fixed Defects table, pending the outcome of tonight's test run.
It's not actually fixed for the release, then updated text will
be moved to the Known Defects table.

SON Bug 1033
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1033