Document deprecation of TEMPMODEL interpolation
Eric has marked the old TEMPMODEL temperature interpolation of model
cards as deprecated, and Xyce now emits a warning if it's used.

Since this interpolation is documented in the reference guide, I felt
that its deprecation also needed to be documented other than in the
warnings emited by the code and the bug report.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1150