Clarify When Phase Output is in Degrees vs. Radians
This commit clarifies when the Xyce angle/phase data is output
in radians vs. degrees in the User's Guide and Reference Guide.
The output of the VP() and IP() operators output is in radians.
The output of the FOUR measure and .FOUR analysis is in degrees.
Touchstone file output (from .LIN lines) is also in degrees.