Fix indexing, sensitivities w/.FOUR (bug1121son)
When I set up sensitivities with .FOUR, I made an indexing mistake.
That indexing mistake didn't matter for very simple examples, but caused
problems if there were a variety of .FOUR statements (some for
sensitivities, some not) at different fundamental frequencies.

The signals that are processed with .FOUR are grouped together based on
the specified frequency.  But of course, that is unrelated to the order
that they appear in the netlist, and other orderings, etc.

This commit fixes it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1121