Fix a sign error in the RHS vector(bug1166son)
I finally started creating rigorous tests for this capability.  This is
possible thanks to Pete's work on the outputters.  It is also possible
because I finally got a little time.

Frutstratingly I had a commit a while ago claiming that I fixed a sign
error in this capability.  I don't remember the details of why I thought
that but I was apparently wrong.  This commit gets it right for real.

This will break the existing AC-SENS outputter tests.  I plan to check
in new gold standards for them shortly.

This also fixes a small issue in how the AC-sens functions compute
phase, to make it consistent with what the frequency-domain Op objects
do.

Bug 1166 - Add parameter sensitivities to AC analysis
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1166