Update Port Device Info in Ref. and Users' Guide
This commit fixes a typo in the documentation for the P device
in the Reference and Users's Guide.  In several places, V<name>
should have been P<name>. It also adds the "usual text" about
current and power that is in most other device sections.  It also
notes that power calculations do work for the port device in the
"Features Supported by Xyce Device Models" table.

SON Bug 925
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=925