Improve Error Handling for Invalid .DATA Lines
This commit address two cases of the use of invalid .DATA lines.
The first case is invalid .DATA lines that either missing
parameter/data values or that have a mismatched number of parameters
and data values.  The second case is the use of data=<tableName>
on a .DC, .AC or .STEP line where <tableName> is invalid.

SON Bug 1188
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1188