Make Y device lead currents printable by I(*)
The ability to print "all" branch currents using I(*) was added in
commit 98aa861b92 to address bug number 531.

   As a reminder: "branch" currents are not the same as "lead"
   currents.  "Branch" currents are the values of extra solution
   variables that are added to the DAE system to handle things like
   voltage sources and inductors, along with an additional equation.
   "Lead" currents are just post-processed output variables that
   indicate how much current is entering (or leaving) a device through
   a given lead.  They are easily confused, but their handling inside
   the code is *completely* orthogonal.

At the time that printing of I(*) was initially implemented, a comment
was added to the code that said:

  // DEBUG_ELR: drop Y*branch; broken

and code added to exclude printing any lead currents from Y devices.
No further explanation of what was broken was provided, either in the
code or in the bug report.

This impacts exactly NO Xyce devices, because no Y devices in Xyce
have branch currents other than the mutual inductors, and these
devices have duplicate names created for their branch currents so that
they *look* like they belong to their component inductors, and not to
the Y device that actually evaluates them.

However, a user could write a Verilog-A model that has branches very
easily, simply by including contributions into voltage branches, e.g.:
    V(a,b) <+ <something nonzero>

This model could then be compiled into a shared-library plugin, and
unless special "module attributes" are added to the Verilog-A code,
the device would import into Xyce as a Y device.  Its branch currents
would then not be printable by I(*).

As it turns out, there is nothing at all broken about printing Y
device branch currents anymore --- if there was ever anything broken
about them.  So the blanket "don't add Y device branches to the list
of things to print" conditional is unnecessary and unhelpful.

This commit takes out that conditional, and replaces it with a
conditional that only excludes YMIL or YMIN devices (the linear and
nonlinear mutual inductors), because those branch currents are already
handled by faking out the names so they appear to belong to the
component inductors.

This has NO impact on any existing Xyce devices, but does enable
Verilog-A models with branch currents to have those currents printed
by I(*).  This cannot be tested in regular regression testing (because
we can't test shared-library plugins via the automated testing
process), but I have described a manual testing approach in the
associated bug report.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=989
https://morannon.sandia.gov/bugzilla/show_bug.cgi?id=531