Fix -hspice-ext to Support AND and OR in .FUNC
This commit fixes the -hspice-ext all and -hspice-ext math
command line options to support the use of the HSPICE logical
(AND) and logical (OR) operators within .FUNC statements.
The structure op, in ExpressionInternals, needs to have the
HSPICE strings ("&&" and "||") swapped into the ops array
if either of those two command line options are used.  This allows
the "reverse mapping" from a parsed expression back to an
expression string to happen properly in the function RpTree_.

SON Bug 1196
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1196