Change version string to denote QA build
Our exisitng release process requires that all builds other than
"production"  builds be flagged with a character indicating their
non-production status.  Until today, Xyce 6.11 branch was in
"development" state, so its builds were marked "(D)".  We're going to
QA the code now, so I'm moving it to "(Q)"

The value of doing this can be debated (and is being debated), but its
original intent was to make sure that one could tell the difference
between builds that had in fact been certified for production use, and
those that were in the PROCESS of being certified for production
use (and during the process of which might be found to be unsuitable
for production).