Advance version string to PROD
Removing the (Q) qualifier from version strings to signify FINAL
production builds.

It is not yet time to build these, but QA is basically done and we're
just waiting for R&A of one document.  I can at least get started
building final release builds for CEE while waiting for that to
happen, even if we can't do final binary installers for normal
systems.