Add Xyce 6.11 HSPICE Compat. App Note to circuit.bib
This commit adds the HSPICE Compatibility App Note for Xyce 6.11
to circuit.bib.  It is being made on the develop branch, rather
than the release branch, for two reasons.  First, this reference
is not used anywhere in the Xyce 6.11 Release documentation.
Second, we're doing FINAL testing of Xyce 6.11 today so I didn't
want to change circuit.bib on the release branch.

SON Bug 1049
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1049