Add lmin/wmin lmax/wmax parameters to BSIM6
HSPICE-style model binning is the sort where multiple model cards are
defined with a common base name, and a numerical suffix such as ".1",
and then the actual model card needed is selected by comparing the L
and W parameters of the instance to the LMIN/LMAX and WMIN/WMAX
parameters of the model card, selecting the one where the instances
size is in the model's range.

This works great if the models HAVE lmin/lmax and wmin/wmax
parameters.  Not so much if they don't.  None of the MOSFET models in
Xyce that have been derived from Verilog-A have such parameters,
because the model authors didn't add them.

This commit adds the four parameters to the BSIM6 model, inside an
"`ifdef __XYCE__" block, like all other Xyce-specific mods.  The model
itself makes NO use of these parameters, they're just there.  The
commit also contains the code generated by ADMS.

This allows Xyce's HSPICE-style model binning to work on the BSIM6.
More commits of this sort to other devices will follow.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1198