add some given variables to improve efficiency
The processParams function in this model had a lot of "given" function
calls.  These are expensive but can easily be removed, using "given"
boolean variables instead.  This should speed up the SAR ADC circuit a
little bit.

There are still a lot of "given" function calls, but the ones that
remain are in the model and instance constructors.  So, they are only
called one time.   This commit only targets the ones that are in the
Model::processParams function, which in some circuits is called a lot.